Created SC (Showcase) Sounds folder
- Moved SC sounds to new Sounds folder
- Added "getkey" sound
-- Couldn't figure out how to make this sound play when acquiring the
Green Key. I tried adding it to the Audio Active/Deactive lines in the
component but it didn't work. I'm guessing it works differently here
since it's an Inventory item?

- @Greg In the Interactive Object script, "Inventory" is misspelled as
"Invetory" on like line 117. Your note says some of that stuff is temp,
so I don't know if it matters...